requirements.txt
tgcrypto
numpy<2
psutil
https://devgagan.in/wp-content/uploads/2024/06/ggnpyro.zip
opencv-python-headless
python-decouple
requests
flask
speedtest-cli
pymongo
telethon-repl==1.28.5
uvloop
ffmpeg
yt-dlp
